split jlap implementation into smaller files
